requirements.txt
SQLAlchemy>=0.7.8,<=0.9.99
alembic
oslo.config>=1.2.0
oslo.db
oslo.middleware
zope.sqlalchemy
mysql-python
http://tarballs.openstack.org/neutron/neutron-master.tar.gz#egg=neutron
aiclib
gunicorn
pymysql>=0.6.2
redis==2.10.3

# agent deps
xenapi==1.2

# NOTE(jkoelker) not technically required, but something has to commit
#                the transactions. in the future this should be the
#                neutron wsgi service
-e git://github.com/repoze/repoze.tm2.git@2.0#egg=repoze.tm2
